refactor: unify chat message persistence into callback pattern
Extract Axiom query helpers to src/lib/run/extract-chat-events.ts,
expand the chat callback consumer to handle the full completion flow
(message persistence, title generation, error messages), and remove
inline chat persistence from the complete webhook.

Closes #7377
