+++ turbo/apps/web/app/api/internal/callbacks/chat/__tests__/route.test.ts
@@ -1,4 +1,5 @@
 import { describe, it, expect, beforeEach, vi } from "vitest";
+import { HttpResponse } from "msw";
 import {
   testContext,
   uniqueId,
@@ -14,8 +15,11 @@ import {
 } from "../../../../../../src/__tests__/api-test-helpers";
 import { computeHmacSignature } from "../../../../../../src/lib/callback/hmac";
 import { reloadEnv } from "../../../../../../src/env";
+import { getSessionChatMessages } from "../../../../../../src/lib/agent-session";
 import { POST as createThreadHandler } from "../../../../zero/chat-threads/route";
 import { POST } from "../route";
+import { http } from "../../../../../../src/__tests__/msw";
+import { server } from "../../../../../../src/mocks/server";
 
 const context = testContext();
 
@@ -58,8 +62,12 @@ describe("POST /api/internal/callbacks/chat", () => {
     reloadEnv();
   });
 
-  /** Create a thread via route handler, then a run and callback in DB. */
-  async function setupRunAndThread() {
+  /** Create a thread via route handler, then a run, session, and callback in DB. */
+  async function setupRunAndThread(
+    options: { status?: "completed" | "failed" } = {},
+  ) {
+    const { status = "completed" } = options;
+
     // Create a chat thread via the route handler
     const threadResponse = await createThreadHandler(
       createTestRequest("http://localhost:3000/api/zero/chat-threads", {
@@ -76,8 +84,11 @@ describe("POST /api/internal/callbacks/chat", () => {
 
     // Create a run in DB
     const { runId } = await createTestRunInDb(user.userId, agentId, {
-      status: "completed",
+      status,
     });
+
+    // Create an agent session so findNewSessionId() returns a session
+    const session = await createTestAgentSession(user.userId, agentId);
 
     // Create a callback record
     const { secret } = await createTestCallback({
@@ -86,7 +97,7 @@ describe("POST /api/internal/callbacks/chat", () => {
       payload: { threadId, agentId },
     });
 
-    return { threadId, runId, secret };
+    return { threadId, runId, secret, sessionId: session.id };
   }
 
   /** Get thread detail and extract latestSessionId. */
@@ -100,6 +111,19 @@ describe("POST /api/internal/callbacks/chat", () => {
     );
     const data = await response.json();
     return data.latestSessionId ?? null;
+  }
+
+  /** Get thread title via the API. */
+  async function getThreadTitle(threadId: string): Promise<string | null> {
+    const { GET } = await import("../../../../zero/chat-threads/[id]/route");
+    const response = await GET(
+      createTestRequest(
+        `http://localhost:3000/api/zero/chat-threads/${threadId}`,
+        { method: "GET" },
+      ),
+    );
+    const data = await response.json();
+    return data.title ?? null;
   }
 
   it("should return 200 for progress status without updating sessionId", async () => {
@@ -125,35 +149,11 @@ describe("POST /api/internal/callbacks/chat", () => {
     expect(sessionId).toBeNull();
   });
 
-  it("should return 200 for failed status without updating sessionId", async () => {
-    const { threadId, runId, secret } = await setupRunAndThread();
-
-    const response = await POST(
-      createCallbackRequest(
-        {
-          runId,
-          status: "failed",
-          error: "Something went wrong",
-          payload: { threadId, agentId },
-        },
-        secret,
-      ),
-    );
-
-    expect(response.status).toBe(200);
-    const data = await response.json();
-    expect(data.success).toBe(true);
-
-    // Verify sessionId is still null
-    const sessionId = await getThreadSessionId(threadId);
-    expect(sessionId).toBeNull();
-  });
-
   it("should update sessionId on completion when a matching session exists", async () => {
-    const { threadId, runId, secret } = await setupRunAndThread();
+    const { threadId, runId, secret, sessionId } = await setupRunAndThread();
 
-    // Create an agent session that would match findNewSessionId()
-    const session = await createTestAgentSession(user.userId, agentId);
+    // Mock Axiom to return empty (no result events)
+    context.mocks.axiom.queryAxiom.mockResolvedValueOnce([]);
 
     const response = await POST(
       createCallbackRequest(
@@ -171,15 +171,16 @@ describe("POST /api/internal/callbacks/chat", () => {
     expect(data.success).toBe(true);
 
     // Verify sessionId was updated via thread detail API
-    const sessionId = await getThreadSessionId(threadId);
-    expect(sessionId).toBe(session.id);
+    const threadSessionId = await getThreadSessionId(threadId);
+    expect(threadSessionId).toBe(sessionId);
   });
 
   it("should be idempotent - calling twice does not break", async () => {
     const { threadId, runId, secret } = await setupRunAndThread();
 
-    // Create a matching session
-    await createTestAgentSession(user.userId, agentId);
+    // Mock Axiom for both calls
+    context.mocks.axiom.queryAxiom.mockResolvedValueOnce([]);
+    context.mocks.axiom.queryAxiom.mockResolvedValueOnce([]);
 
     const makeRequest = () =>
       createCallbackRequest(
@@ -223,5 +224,415 @@ describe("POST /api/internal/callbacks/chat", () => {
     );
 
     expect(response.status).toBe(400);
+  });
+
+  describe("Chat Persistence", () => {
+    it("should persist user + assistant messages with summaries on completion", async () => {
+      const { threadId, runId, secret, sessionId } = await setupRunAndThread();
+
+      // Mock Axiom to return assistant events + result event
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
+        {
+          eventType: "assistant",
+          eventData: {
+            message: {
+              content: [{ type: "tool_use", name: "Bash" }],
+            },
+          },
+        },
+        {
+          eventType: "assistant",
+          eventData: {
+            message: {
+              content: [{ type: "tool_use", name: "Read" }],
+            },
+          },
+        },
+        {
+          eventType: "assistant",
+          eventData: {
+            message: {
+              content: [{ type: "text", text: "Done. Created 3 files." }],
+            },
+          },
+        },
+        {
+          eventType: "result",
+          eventData: { result: "Done. Created 3 files." },
+        },
+      ]);
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      // Verify session now has chat messages
+      type StoredMessage = {
+        role: string;
+        content: string;
+        runId?: string;
+        summaries?: Array<
+          { kind: "tool"; name: string } | { kind: "text"; text: string }
+        >;
+      };
+      const chatMessages = (await getSessionChatMessages(
+        sessionId,
+      )) as StoredMessage[];
+      expect(chatMessages).toHaveLength(2);
+
+      // Verify user message from prompt
+      const userMsg = chatMessages.find((m) => m.role === "user");
+      expect(userMsg).toBeDefined();
+      expect(userMsg!.content).toBe("test prompt");
+
+      // Verify assistant message from Axiom result
+      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
+      expect(assistantMsg).toBeDefined();
+      expect(assistantMsg!.content).toBe("Done. Created 3 files.");
+      expect(assistantMsg!.runId).toBe(runId);
+      // Last text event is skipped — only tool_use entries are extracted
+      expect(assistantMsg!.summaries).toEqual([
+        { kind: "tool", name: "Bash" },
+        { kind: "tool", name: "Read" },
+      ]);
+    });
+
+    it("should persist only user message when no result found in Axiom", async () => {
+      const { threadId, runId, secret, sessionId } = await setupRunAndThread();
+
+      // Axiom returns no events
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([]);
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      // Verify session has only user message
+      type StoredMessage = { role: string; content: string };
+      const chatMessages = (await getSessionChatMessages(
+        sessionId,
+      )) as StoredMessage[];
+      expect(chatMessages).toHaveLength(1);
+      expect(chatMessages[0]!.role).toBe("user");
+      expect(chatMessages[0]!.content).toBe("test prompt");
+    });
+
+    it("should not include summaries when Axiom returns no assistant events", async () => {
+      const { threadId, runId, secret, sessionId } = await setupRunAndThread();
+
+      // Combined Axiom query returns result event only (no assistant events)
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
+        {
+          eventType: "result",
+          eventData: { result: "All good." },
+        },
+      ]);
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      type StoredMessage = {
+        role: string;
+        content: string;
+        summaries?: string[];
+      };
+      const chatMessages = (await getSessionChatMessages(
+        sessionId,
+      )) as StoredMessage[];
+
+      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
+      expect(assistantMsg).toBeDefined();
+      expect(assistantMsg!.content).toBe("All good.");
+      // summaries should be undefined (not included when empty)
+      expect(assistantMsg!.summaries).toBeUndefined();
+    });
+
+    it("should truncate text summaries exceeding 80 characters", async () => {
+      const { threadId, runId, secret, sessionId } = await setupRunAndThread();
+
+      const longText = "x".repeat(100);
+
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
+        {
+          eventType: "assistant",
+          eventData: {
+            message: {
+              content: [{ type: "text", text: longText }],
+            },
+          },
+        },
+        {
+          eventType: "assistant",
+          eventData: {
+            message: {
+              content: [{ type: "text", text: "Done." }],
+            },
+          },
+        },
+        {
+          eventType: "result",
+          eventData: { result: "Done." },
+        },
+      ]);
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      type StoredMessage = {
+        role: string;
+        summaries?: Array<
+          { kind: "tool"; name: string } | { kind: "text"; text: string }
+        >;
+      };
+      const chatMessages = (await getSessionChatMessages(
+        sessionId,
+      )) as StoredMessage[];
+
+      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
+      expect(assistantMsg).toBeDefined();
+      expect(assistantMsg!.summaries).toHaveLength(1);
+      const entry = assistantMsg!.summaries![0] as {
+        kind: "text";
+        text: string;
+      };
+      expect(entry.kind).toBe("text");
+      expect(entry.text.length).toBe(81); // 80 chars + "…"
+      expect(entry.text.endsWith("\u2026")).toBe(true);
+    });
+
+    it("should persist user + error messages on failed run", async () => {
+      const { threadId, runId, secret, sessionId } = await setupRunAndThread({
+        status: "failed",
+      });
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "failed",
+            error: "Agent crashed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      type StoredMessage = {
+        role: string;
+        content: string;
+        runId?: string;
+      };
+      const chatMessages = (await getSessionChatMessages(
+        sessionId,
+      )) as StoredMessage[];
+      expect(chatMessages).toHaveLength(2);
+
+      const userMsg = chatMessages.find((m) => m.role === "user");
+      expect(userMsg).toBeDefined();
+      expect(userMsg!.content).toBe("test prompt");
+
+      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
+      expect(assistantMsg).toBeDefined();
+      expect(assistantMsg!.content).toBe("Agent crashed");
+      expect(assistantMsg!.runId).toBe(runId);
+    });
+
+    it("should update sessionId on failed run", async () => {
+      const { threadId, runId, secret, sessionId } = await setupRunAndThread({
+        status: "failed",
+      });
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "failed",
+            error: "Agent crashed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      const threadSessionId = await getThreadSessionId(threadId);
+      expect(threadSessionId).toBe(sessionId);
+    });
+  });
+
+  describe("Title Generation", () => {
+    function mockOpenRouter(title: string) {
+      const { handler, mocked } = http.post(
+        "https://openrouter.ai/api/v1/chat/completions",
+        () =>
+          HttpResponse.json({
+            choices: [{ message: { content: title } }],
+          }),
+      );
+      server.use(handler);
+      return mocked;
+    }
+
+    function mockOpenRouterError(status: number) {
+      const { handler, mocked } = http.post(
+        "https://openrouter.ai/api/v1/chat/completions",
+        () => new HttpResponse("Internal Server Error", { status }),
+      );
+      server.use(handler);
+      return mocked;
+    }
+
+    it("should generate and set chat thread title on completion", async () => {
+      const { threadId, runId, secret } = await setupRunAndThread();
+
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
+        {
+          eventType: "result",
+          eventData: { result: "Use --inspect flag for debugging." },
+        },
+      ]);
+
+      vi.stubEnv("OPENROUTER_API_KEY", "test-openrouter-key");
+      reloadEnv();
+      const openRouterMock = mockOpenRouter("Debugging Node.js Apps");
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+      expect(openRouterMock).toHaveBeenCalledTimes(1);
+
+      const title = await getThreadTitle(threadId);
+      expect(title).toBe("Debugging Node.js Apps");
+    });
+
+    it("should not generate title on failed run", async () => {
+      const { threadId, runId, secret } = await setupRunAndThread({
+        status: "failed",
+      });
+
+      vi.stubEnv("OPENROUTER_API_KEY", "test-openrouter-key");
+      reloadEnv();
+      const openRouterMock = mockOpenRouter("Should not be called");
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "failed",
+            error: "Agent crashed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+      expect(openRouterMock).not.toHaveBeenCalled();
+    });
+
+    it("should not fail callback when OpenRouter returns an error", async () => {
+      const { threadId, runId, secret } = await setupRunAndThread();
+
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
+        { eventType: "result", eventData: { result: "Some result" } },
+      ]);
+
+      vi.stubEnv("OPENROUTER_API_KEY", "test-openrouter-key");
+      reloadEnv();
+      mockOpenRouterError(500);
+
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      // Thread title should remain unchanged (error was caught)
+      const title = await getThreadTitle(threadId);
+      expect(title).toBe("Test thread");
+    });
+
+    it("should skip title generation when OPENROUTER_API_KEY is not set", async () => {
+      const { threadId, runId, secret } = await setupRunAndThread();
+
+      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
+        { eventType: "result", eventData: { result: "Some result" } },
+      ]);
+
+      // Do NOT set OPENROUTER_API_KEY — feature should be a no-op
+      const response = await POST(
+        createCallbackRequest(
+          {
+            runId,
+            status: "completed",
+            payload: { threadId, agentId },
+          },
+          secret,
+        ),
+      );
+
+      expect(response.status).toBe(200);
+
+      // Thread title should remain as initial value
+      const title = await getThreadTitle(threadId);
+      expect(title).toBe("Test thread");
+    });
   });
 });
