+++ turbo/apps/web/app/api/internal/callbacks/chat/route.ts
@@ -1,10 +1,15 @@
 import { NextRequest, NextResponse } from "next/server";
 import { eq } from "drizzle-orm";
+import type { SummaryEntry } from "@vm0/core";
 import { initServices } from "../../../../../src/lib/init-services";
 import { verifyCallback } from "../../../../../src/lib/callback";
 import { chatThreads } from "../../../../../src/db/schema/chat-thread";
 import { agentRuns } from "../../../../../src/db/schema/agent-run";
 import { findNewSessionId } from "../../../../../src/lib/session/find-new-session";
+import { queryRunEventsForChat } from "../../../../../src/lib/run/extract-chat-events";
+import { appendChatMessages } from "../../../../../src/lib/agent-session/agent-session-service";
+import { generateChatTitle } from "../../../../../src/lib/ai/lightweight-model";
+import { updateChatThreadTitle } from "../../../../../src/lib/chat-thread";
 import type { ChatCallbackPayload } from "../../../../../src/lib/callback/callback-payloads";
 import { logger } from "../../../../../src/lib/logger";
 
@@ -20,10 +25,136 @@ function parsePayload(payload: unknown): ChatCallbackPayload | null {
 }
 
 /**
+ * Persist chat messages to the session and update thread sessionId if needed.
+ */
+async function persistMessages(
+  sessionId: string,
+  userId: string,
+  messages: Array<{
+    role: "user" | "assistant";
+    content: string;
+    runId?: string;
+    summaries?: SummaryEntry[];
+  }>,
+  runId: string,
+): Promise<void> {
+  try {
+    await appendChatMessages(sessionId, userId, messages);
+    log.debug(`Persisted ${messages.length} chat messages for run ${runId}`);
+  } catch (err) {
+    log.error("Failed to persist chat messages", { err });
+  }
+}
+
+/**
+ * Update sessionId on thread if not already set.
+ */
+async function updateThreadSessionId(
+  threadId: string,
+  sessionId: string,
+): Promise<void> {
+  const [thread] = await globalThis.services.db
+    .select({ sessionId: chatThreads.sessionId })
+    .from(chatThreads)
+    .where(eq(chatThreads.id, threadId))
+    .limit(1);
+
+  if (thread && !thread.sessionId) {
+    await globalThis.services.db
+      .update(chatThreads)
+      .set({ sessionId })
+      .where(eq(chatThreads.id, threadId));
+  }
+}
+
+/**
+ * Handle completed run: persist messages with summaries, update sessionId, generate title.
+ */
+async function handleCompleted(
+  runId: string,
+  sessionId: string | undefined,
+  userId: string,
+  prompt: string,
+  threadId: string,
+): Promise<void> {
+  // Query Axiom for result text and summaries
+  let resultText: string | null = null;
+  let summaries: SummaryEntry[] = [];
+  try {
+    const axiomResult = await queryRunEventsForChat(runId);
+    resultText = axiomResult.resultText;
+    summaries = axiomResult.summaries;
+  } catch (err) {
+    log.error("Failed to query Axiom for chat events", { err });
+  }
+
+  // Persist chat messages to session
+  if (sessionId) {
+    const messages: Array<{
+      role: "user" | "assistant";
+      content: string;
+      runId?: string;
+      summaries?: SummaryEntry[];
+    }> = [{ role: "user", content: prompt }];
+
+    if (resultText) {
+      messages.push({
+        role: "assistant",
+        content: resultText,
+        runId,
+        ...(summaries.length > 0 ? { summaries } : {}),
+      });
+    }
+
+    await persistMessages(sessionId, userId, messages, runId);
+    await updateThreadSessionId(threadId, sessionId);
+  }
+
+  // Generate and update chat thread title
+  try {
+    const title = await generateChatTitle(prompt, resultText);
+    if (title) {
+      await updateChatThreadTitle(threadId, title);
+    }
+  } catch (err) {
+    log.warn("Failed to generate chat title", { err });
+  }
+}
+
+/**
+ * Handle failed run: persist error messages, update sessionId.
+ */
+async function handleFailed(
+  runId: string,
+  sessionId: string | undefined,
+  userId: string,
+  prompt: string,
+  threadId: string,
+  errorMessage: string,
+): Promise<void> {
+  if (sessionId) {
+    await persistMessages(
+      sessionId,
+      userId,
+      [
+        { role: "user", content: prompt },
+        { role: "assistant", content: errorMessage, runId },
+      ],
+      runId,
+    );
+    await updateThreadSessionId(threadId, sessionId);
+  }
+}
+
+/**
  * POST /api/internal/callbacks/chat
  *
  * Chat callback handler for agent run completion.
- * Persists sessionId on the chat thread when a run completes.
+ * Handles the full chat completion flow:
+ * - Persists user + assistant messages (with summaries) to the session
+ * - Sets sessionId on the chat thread
+ * - Generates and updates the chat thread title (on completion only)
+ * - Persists error messages on failure
  */
 export async function POST(request: NextRequest): Promise<NextResponse> {
   initServices();
@@ -31,7 +162,7 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
   const result = await verifyCallback<ChatCallbackPayload>(request, log);
   if (!result.ok) return result.response;
 
-  const { runId, status } = result.data;
+  const { runId, status, error } = result.data;
   const payload = parsePayload(result.data.payload);
   if (!payload) {
     return NextResponse.json(
@@ -40,14 +171,19 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
     );
   }
 
-  // Progress and failure: no session update needed
-  if (status !== "completed") {
+  // Progress: no-op
+  if (status === "progress") {
     return NextResponse.json({ success: true });
   }
 
-  // On completion: persist sessionId on the chat thread
+  // Fetch the run record for userId, prompt, createdAt, error
   const [run] = await globalThis.services.db
-    .select({ userId: agentRuns.userId, createdAt: agentRuns.createdAt })
+    .select({
+      userId: agentRuns.userId,
+      prompt: agentRuns.prompt,
+      createdAt: agentRuns.createdAt,
+      error: agentRuns.error,
+    })
     .from(agentRuns)
     .where(eq(agentRuns.id, runId))
     .limit(1);
@@ -56,25 +192,31 @@ export async function POST(request: NextRequest): Promise<NextResponse> {
     return NextResponse.json({ success: true });
   }
 
-  // Check if thread already has a sessionId
-  const [thread] = await globalThis.services.db
-    .select({ sessionId: chatThreads.sessionId })
-    .from(chatThreads)
-    .where(eq(chatThreads.id, payload.threadId))
-    .limit(1);
+  // Resolve session for message persistence
+  const sessionId = await findNewSessionId(
+    run.userId,
+    payload.agentId,
+    run.createdAt,
+  );
 
-  if (thread && !thread.sessionId) {
-    const newSessionId = await findNewSessionId(
+  if (status === "completed") {
+    await handleCompleted(
+      runId,
+      sessionId,
       run.userId,
-      payload.agentId,
-      run.createdAt,
+      run.prompt,
+      payload.threadId,
     );
-    if (newSessionId) {
-      await globalThis.services.db
-        .update(chatThreads)
-        .set({ sessionId: newSessionId })
-        .where(eq(chatThreads.id, payload.threadId));
-    }
+  } else if (status === "failed") {
+    const errorMessage = error ?? run.error ?? "Run failed";
+    await handleFailed(
+      runId,
+      sessionId,
+      run.userId,
+      run.prompt,
+      payload.threadId,
+      errorMessage,
+    );
   }
 
   return NextResponse.json({ success: true });
