+++ turbo/apps/web/app/api/webhooks/agent/complete/__tests__/route.test.ts
@@ -18,10 +18,7 @@ import {
   findTestQueueEntry,
 } from "../../../../../../src/__tests__/api-test-helpers";
 import { reloadEnv } from "../../../../../../src/env";
-import {
-  getAgentSessionWithConversation,
-  getSessionChatMessages,
-} from "../../../../../../src/lib/agent-session";
+import { getAgentSessionWithConversation } from "../../../../../../src/lib/agent-session";
 import {
   testContext,
   uniqueId,
@@ -399,514 +396,6 @@ describe("POST /api/webhooks/agent/complete", () => {
       const data = await response.json();
       expect(data.success).toBe(true);
       expect(data.status).toBe("failed");
-    });
-  });
-
-  describe("Chat Persistence", () => {
-    it("should persist chat messages to session after successful completion", async () => {
-      // Create checkpoint (which creates a session)
-      const checkpointRequest = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/checkpoints",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            cliAgentType: "claude-code",
-            cliAgentSessionId: "test-session-chat",
-            cliAgentSessionHistory: JSON.stringify({ type: "test" }),
-            artifactSnapshot: {
-              artifactName: "test-artifact",
-              artifactVersion: "v1",
-            },
-          }),
-        },
-      );
-      const checkpointResponse = await checkpointWebhook(checkpointRequest);
-      expect(checkpointResponse.status).toBe(200);
-      const checkpointData = (await checkpointResponse.json()) as {
-        agentSessionId: string;
-      };
-
-      // Mock Axiom to return a result event for this run
-      // persistChatMessages uses a single combined query for both result and assistant events
-      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
-        {
-          eventType: "result",
-          eventData: { result: "Here is the agent response." },
-        },
-      ]);
-
-      // Complete the run
-      const request = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/complete",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            exitCode: 0,
-          }),
-        },
-      );
-
-      const response = await POST(request);
-      expect(response.status).toBe(200);
-
-      // Flush after() callbacks which trigger persistChatMessages
-      await context.mocks.flushAfter();
-
-      // Verify session now has chat messages
-      type StoredMessage = { role: string; content: string; runId?: string };
-      const chatMessages = (await getSessionChatMessages(
-        checkpointData.agentSessionId,
-      )) as StoredMessage[];
-      expect(chatMessages.length).toBeGreaterThanOrEqual(2);
-
-      // Verify user message from prompt
-      const userMsg = chatMessages.find((m) => m.role === "user");
-      expect(userMsg).toBeDefined();
-      expect(userMsg!.content).toBe("Test prompt");
-
-      // Verify assistant message from Axiom result
-      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
-      expect(assistantMsg).toBeDefined();
-      expect(assistantMsg!.content).toBe("Here is the agent response.");
-      expect(assistantMsg!.runId).toBe(testRunId);
-    });
-
-    it("should persist only user message when no result found in Axiom", async () => {
-      // Create checkpoint
-      const checkpointRequest = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/checkpoints",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            cliAgentType: "claude-code",
-            cliAgentSessionId: "test-session-no-result",
-            cliAgentSessionHistory: JSON.stringify({ type: "test" }),
-            artifactSnapshot: {
-              artifactName: "test-artifact",
-              artifactVersion: "v1",
-            },
-          }),
-        },
-      );
-      const checkpointResponse = await checkpointWebhook(checkpointRequest);
-      expect(checkpointResponse.status).toBe(200);
-      const checkpointData = (await checkpointResponse.json()) as {
-        agentSessionId: string;
-      };
-
-      // Axiom returns no events (default mock)
-      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([]);
-
-      // Complete the run
-      const request = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/complete",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            exitCode: 0,
-          }),
-        },
-      );
-
-      const response = await POST(request);
-      expect(response.status).toBe(200);
-
-      await context.mocks.flushAfter();
-
-      // Verify session has only user message
-      type StoredMessage = { role: string; content: string };
-      const chatMessages = (await getSessionChatMessages(
-        checkpointData.agentSessionId,
-      )) as StoredMessage[];
-      expect(chatMessages).toHaveLength(1);
-      expect(chatMessages[0]!.role).toBe("user");
-      expect(chatMessages[0]!.content).toBe("Test prompt");
-    });
-
-    it("should persist summaries extracted from Axiom events", async () => {
-      // Create checkpoint (which creates a session)
-      const checkpointRequest = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/checkpoints",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            cliAgentType: "claude-code",
-            cliAgentSessionId: "test-session-summaries",
-            cliAgentSessionHistory: JSON.stringify({ type: "test" }),
-            artifactSnapshot: {
-              artifactName: "test-artifact",
-              artifactVersion: "v1",
-            },
-          }),
-        },
-      );
-      const checkpointResponse = await checkpointWebhook(checkpointRequest);
-      expect(checkpointResponse.status).toBe(200);
-      const checkpointData = (await checkpointResponse.json()) as {
-        agentSessionId: string;
-      };
-
-      // Combined Axiom query returns both result and assistant events
-      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "tool_use", name: "Bash" }],
-            },
-          },
-        },
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "tool_use", name: "Read" }],
-            },
-          },
-        },
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "text", text: "Done. Created 3 files." }],
-            },
-          },
-        },
-        {
-          eventType: "result",
-          eventData: { result: "Done. Created 3 files." },
-        },
-      ]);
-
-      // Complete the run
-      const request = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/complete",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            exitCode: 0,
-          }),
-        },
-      );
-
-      const response = await POST(request);
-      expect(response.status).toBe(200);
-
-      await context.mocks.flushAfter();
-
-      // Verify summaries are persisted on the assistant message
-      type StoredMessage = {
-        role: string;
-        content: string;
-        runId?: string;
-        summaries?: Array<
-          | string
-          | { kind: "tool"; name: string }
-          | { kind: "text"; text: string }
-        >;
-      };
-      const chatMessages = (await getSessionChatMessages(
-        checkpointData.agentSessionId,
-      )) as StoredMessage[];
-      expect(chatMessages.length).toBeGreaterThanOrEqual(2);
-
-      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
-      expect(assistantMsg).toBeDefined();
-      expect(assistantMsg!.content).toBe("Done. Created 3 files.");
-      // Last text event is skipped — only tool_use entries are extracted
-      expect(assistantMsg!.summaries).toEqual([
-        { kind: "tool", name: "Bash" },
-        { kind: "tool", name: "Read" },
-      ]);
-    });
-
-    it("should extract text summaries when events have no tool_use", async () => {
-      const checkpointRequest = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/checkpoints",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            cliAgentType: "claude-code",
-            cliAgentSessionId: "test-session-text-summaries",
-            cliAgentSessionHistory: JSON.stringify({ type: "test" }),
-            artifactSnapshot: {
-              artifactName: "test-artifact",
-              artifactVersion: "v1",
-            },
-          }),
-        },
-      );
-      const checkpointResponse = await checkpointWebhook(checkpointRequest);
-      expect(checkpointResponse.status).toBe(200);
-      const checkpointData = (await checkpointResponse.json()) as {
-        agentSessionId: string;
-      };
-
-      // Combined Axiom query returns both result and assistant events
-      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "text", text: "Let me check the logs first" }],
-            },
-          },
-        },
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "text", text: "Analysis complete." }],
-            },
-          },
-        },
-        {
-          eventType: "result",
-          eventData: { result: "Analysis complete." },
-        },
-      ]);
-
-      const request = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/complete",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            exitCode: 0,
-          }),
-        },
-      );
-
-      const response = await POST(request);
-      expect(response.status).toBe(200);
-
-      await context.mocks.flushAfter();
-
-      type StoredMessage = {
-        role: string;
-        summaries?: Array<
-          | string
-          | { kind: "tool"; name: string }
-          | { kind: "text"; text: string }
-        >;
-      };
-      const chatMessages = (await getSessionChatMessages(
-        checkpointData.agentSessionId,
-      )) as StoredMessage[];
-
-      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
-      expect(assistantMsg).toBeDefined();
-      // First text event included, last text event skipped
-      expect(assistantMsg!.summaries).toEqual([
-        { kind: "text", text: "Let me check the logs first" },
-      ]);
-    });
-
-    it("should truncate text summaries exceeding 80 characters", async () => {
-      const checkpointRequest = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/checkpoints",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            cliAgentType: "claude-code",
-            cliAgentSessionId: "test-session-truncate",
-            cliAgentSessionHistory: JSON.stringify({ type: "test" }),
-            artifactSnapshot: {
-              artifactName: "test-artifact",
-              artifactVersion: "v1",
-            },
-          }),
-        },
-      );
-      const checkpointResponse = await checkpointWebhook(checkpointRequest);
-      expect(checkpointResponse.status).toBe(200);
-      const checkpointData = (await checkpointResponse.json()) as {
-        agentSessionId: string;
-      };
-
-      const longText = "x".repeat(100);
-
-      // Combined Axiom query returns both result and assistant events
-      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "text", text: longText }],
-            },
-          },
-        },
-        {
-          eventType: "assistant",
-          eventData: {
-            message: {
-              content: [{ type: "text", text: "Done." }],
-            },
-          },
-        },
-        {
-          eventType: "result",
-          eventData: { result: "Done." },
-        },
-      ]);
-
-      const request = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/complete",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            exitCode: 0,
-          }),
-        },
-      );
-
-      const response = await POST(request);
-      expect(response.status).toBe(200);
-
-      await context.mocks.flushAfter();
-
-      type StoredMessage = {
-        role: string;
-        summaries?: Array<
-          | string
-          | { kind: "tool"; name: string }
-          | { kind: "text"; text: string }
-        >;
-      };
-      const chatMessages = (await getSessionChatMessages(
-        checkpointData.agentSessionId,
-      )) as StoredMessage[];
-
-      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
-      expect(assistantMsg).toBeDefined();
-      expect(assistantMsg!.summaries).toHaveLength(1);
-      const entry = assistantMsg!.summaries![0] as {
-        kind: "text";
-        text: string;
-      };
-      expect(entry.kind).toBe("text");
-      expect(entry.text.length).toBe(81); // 80 chars + "…"
-      expect(entry.text.endsWith("…")).toBe(true);
-    });
-
-    it("should not include summaries when Axiom returns no message events", async () => {
-      const checkpointRequest = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/checkpoints",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            cliAgentType: "claude-code",
-            cliAgentSessionId: "test-session-no-summaries",
-            cliAgentSessionHistory: JSON.stringify({ type: "test" }),
-            artifactSnapshot: {
-              artifactName: "test-artifact",
-              artifactVersion: "v1",
-            },
-          }),
-        },
-      );
-      const checkpointResponse = await checkpointWebhook(checkpointRequest);
-      expect(checkpointResponse.status).toBe(200);
-      const checkpointData = (await checkpointResponse.json()) as {
-        agentSessionId: string;
-      };
-
-      // Combined Axiom query returns result event only (no assistant events)
-      context.mocks.axiom.queryAxiom.mockResolvedValueOnce([
-        {
-          eventType: "result",
-          eventData: { result: "All good." },
-        },
-      ]);
-
-      const request = createTestRequest(
-        "http://localhost:3000/api/webhooks/agent/complete",
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-            Authorization: `Bearer ${testToken}`,
-          },
-          body: JSON.stringify({
-            runId: testRunId,
-            exitCode: 0,
-          }),
-        },
-      );
-
-      const response = await POST(request);
-      expect(response.status).toBe(200);
-
-      await context.mocks.flushAfter();
-
-      type StoredMessage = {
-        role: string;
-        content: string;
-        summaries?: string[];
-      };
-      const chatMessages = (await getSessionChatMessages(
-        checkpointData.agentSessionId,
-      )) as StoredMessage[];
-
-      const assistantMsg = chatMessages.find((m) => m.role === "assistant");
-      expect(assistantMsg).toBeDefined();
-      expect(assistantMsg!.content).toBe("All good.");
-      // summaries should be undefined (not included when empty)
-      expect(assistantMsg!.summaries).toBeUndefined();
     });
   });
