+++ turbo/apps/web/app/api/webhooks/agent/complete/route.ts
@@ -23,152 +23,9 @@ import { logger } from "../../../../../src/lib/logger";
 import { drainOrgQueue } from "../../../../../src/lib/run/run-queue-service";
 import { dispatchQueuedZeroRun } from "../../../../../src/lib/zero/zero-queue-service";
 import { processOrgCredits } from "../../../../../src/lib/credit/credit-service";
-import { appendChatMessages } from "../../../../../src/lib/agent-session/agent-session-service";
-import {
-  queryAxiom,
-  getDatasetName,
-  DATASETS,
-} from "../../../../../src/lib/axiom";
 import { after } from "next/server";
-import type { SummaryEntry } from "@vm0/core";
 
 const log = logger("webhook:complete");
-
-interface AxiomEventContent {
-  type: string;
-  text?: string;
-  name?: string;
-  input?: Record<string, unknown>;
-}
-
-function summarizeContentBlock(
-  block: AxiomEventContent,
-  skipText: boolean,
-): SummaryEntry | null {
-  if (block.type === "tool_use" && block.name) {
-    return {
-      kind: "tool",
-      name: block.name,
-      ...(block.input ? { input: block.input } : {}),
-    };
-  }
-  if (!skipText && block.type === "text" && block.text) {
-    const line = block.text.split("\n")[0] ?? "";
-    return {
-      kind: "text",
-      text: line.length > 80 ? line.slice(0, 80) + "…" : line,
-    };
-  }
-  return null;
-}
-
-function findLastTextEventIndex(
-  events: Array<{ eventData: { message?: { content?: AxiomEventContent[] } } }>,
-): number {
-  for (let i = events.length - 1; i >= 0; i--) {
-    const content = events[i]?.eventData?.message?.content ?? [];
-    if (content.some((b) => b.type === "text" && b.text)) {
-      return i;
-    }
-  }
-  return -1;
-}
-
-function extractSummariesFromEvents(
-  events: Array<{
-    eventData: { message?: { content?: AxiomEventContent[] } };
-  }>,
-): SummaryEntry[] {
-  const lastTextIdx = findLastTextEventIndex(events);
-  const summaries: SummaryEntry[] = [];
-
-  for (let i = 0; i < events.length; i++) {
-    const content = events[i]?.eventData?.message?.content ?? [];
-    for (const block of content) {
-      const entry = summarizeContentBlock(block, i === lastTextIdx);
-      if (entry) {
-        summaries.push(entry);
-        break;
-      }
-    }
-  }
-  return summaries;
-}
-
-interface CombinedRunEvent {
-  eventType: string;
-  eventData: {
-    result?: string;
-    message?: { content?: AxiomEventContent[] };
-  };
-}
-
-/**
- * Single Axiom query to fetch both "result" and "assistant" events for a run.
- * Replaces two separate queries (extractRunOutput + extractSummariesFromAxiom)
- * to halve the API call count per completion.
- */
-async function queryRunEventsForChat(runId: string): Promise<{
-  resultText: string | null;
-  summaries: SummaryEntry[];
-}> {
-  const dataset = getDatasetName(DATASETS.AGENT_RUN_EVENTS);
-  const apl = `['${dataset}']
-| where runId == "${runId}"
-| where eventType in ("result", "assistant")
-| order by sequenceNumber asc
-| limit 201`; // 200 assistant events + 1 result event
-
-  const events = await queryAxiom<CombinedRunEvent>(apl);
-
-  // Extract last result event
-  const resultEvents = events.filter((e) => e.eventType === "result");
-  const lastResult = resultEvents[resultEvents.length - 1];
-  const resultText =
-    typeof lastResult?.eventData?.result === "string"
-      ? lastResult.eventData.result
-      : null;
-
-  // Extract summaries from assistant events
-  const assistantEvents = events.filter((e) => e.eventType === "assistant");
-  const summaries = extractSummariesFromEvents(assistantEvents);
-
-  return { resultText, summaries };
-}
-
-/**
- * Persist user prompt and assistant result as chat messages on the session.
- * Runs in after() — best-effort, errors are logged but not propagated.
- *
- * Returns the assistant result text (if any) so callers can reuse it.
- */
-async function persistChatMessages(
-  runId: string,
-  sessionId: string,
-  userId: string,
-  prompt: string,
-): Promise<void> {
-  const { resultText, summaries } = await queryRunEventsForChat(runId);
-
-  const messages: Array<{
-    role: "user" | "assistant";
-    content: string;
-    runId?: string;
-    summaries?: SummaryEntry[];
-  }> = [{ role: "user", content: prompt }];
-
-  if (resultText) {
-    messages.push({
-      role: "assistant",
-      content: resultText,
-      runId,
-      ...(summaries.length > 0 ? { summaries } : {}),
-    });
-  }
-
-  await appendChatMessages(sessionId, userId, messages);
-  log.debug(`Persisted ${messages.length} chat messages for run ${runId}`);
-}
 
 /**
  * Schedule terminal side effects in a non-blocking after() block.
@@ -352,20 +209,6 @@ const router = tsr.router(webhookCompleteContract, {
 
       finalStatus = "completed";
       log.debug(`Run ${body.runId} completed successfully`);
-
-      // Persist chat messages (non-blocking)
-      if (session) {
-        after(async () => {
-          await persistChatMessages(
-            body.runId,
-            session.id,
-            userId,
-            run.prompt,
-          ).catch((err) => {
-            log.error("Failed to persist chat messages", { err });
-          });
-        });
-      }
     } else {
       // Failure: store error in run table
       const errorMessage =
